Splitting up pull service and reporting service tests
Makes it easier to group and isolate tests
