+++ test/client/Tug.Client-tests/ClassicPullServerProtocolCompatibilityTests.cs
@@ -1,71 +1,17 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Net.Http;
 using System.Reflection;
 using System.Text;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Logging;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Newtonsoft.Json;
-using Tug.Client.Configuration;
 using Tug.UnitTesting;
 
 namespace Tug.Client
 {
     [TestClass]
-    public class ClassicPullServerProtocolCompatibilityTests
+    public class ClassicPullServerProtocolCompatibilityTests : ProtocolCompatibilityTestsBase
     {
-        // This defines configuration settings for running the
-        // the tests that can be overridden on the command line
-        // by using the format:   /cfg_prop=value
-        public class TestConfig
-        {
-            public string agent_id
-            { get; set; } = "12345678-0000-0000-0000-000000000001";
-
-            public string server_url
-            { get; set; } = "http://DSC-SERVER1.tugnet:8080/PSDSCPullServer.svc/"; // "http://localhost:5000/"; // "http://DSC-LOCALHOST:5000/"; // 
-
-            public string reg_key
-            { get; set; } = "c3ea5066-ce5a-4d12-a42a-850be287b2d8";
-
-            // Only for debugging/testing in DEV (i.e. with Fiddler) -- can't be const because of compile warning
-            public string proxy_url
-            { get; set; } = null; // "http://localhost:8888"; // 
-
-            // Flag that enables some kludges for latest Classic DSC Pull Server on WMF 5.0 platforms
-            //    See:  https://github.com/PowerShell/PowerShell/issues/2921
-            // We leave the default enabled to support the CI server builds, but disable on
-            // local development environments where we're normally testing against Win2016
-            public bool adjust_for_wmf_50
-            { get; set; } = true;
-        }
-
-        // This is the date format that appears to be what the Classic DSC Pull Server
-        // is using to parse and store the report dates, so we need to test against this
-        public const string CLASSIC_SERVER_REPORT_DATE_FORMAT = "yyyy'-'MM'-'dd'T'HH':'mm':'ss'.'fff";
-        // On WMF 5.0, DscService formats its dates differently
-        public const string CLASSIC_SERVER_REPORT_DATE_FORMAT_ALT = "MM/dd/yyyy HH:mm:ss";
-
-        private static TestConfig _testConfig = new TestConfig();
-
-        [ClassInitialize]
-        public static void SetClientLogLevel(TestContext ctx)
-        {
-            // Set the global logging configuration to log some info
-            // which may be useful in debugging and diagnostics
-            Tug.Client.AppLog.Factory.AddConsole(LogLevel.Information);
-
-            // Bind optional configuration overrides to our test config
-            new ConfigurationBuilder()
-                .AddCommandLine(Environment.GetCommandLineArgs()
-                        .Where(x => x.StartsWith("/")).ToArray())
-                .Build()
-                .Bind(_testConfig);
-        }
-
         [TestMethod]
         public void TestRegisterDscAgent()
         {
@@ -454,265 +400,6 @@ namespace Tug.Client
             }
         }
 
-        [TestMethod]
-        public void TestSendReport()
-        {
-            var config = BuildConfig(newAgentId: true);
-            using (var client = new DscPullClient(config))
-            {
-                client.DisableReportAdditionalData = _testConfig.adjust_for_wmf_50;
-                client.RegisterDscAgent().Wait();
-                client.SendReport("SimpleInventoryDefaults",
-                        overrides: new Model.SendReportBody
-                        {
-                            NodeName = "MY_NAME",
-                            IpAddress = "::1;127.0.01",
-                        }).Wait();
-                client.SendReport("DetailedStatusDefaults",
-                        statusData: new[] { "STATUS" }).Wait();
-                client.SendReport("ErrorDefaults",
-                        errors: new[] { "ERROR" }).Wait();
-            }
-        }
-
-        [TestMethod]
-        public void TestSendReport_BadEmptyBody()
-        {
-            var config = BuildConfig();
-            using (var client = new DscPullClient(config))
-            {
-                client.DisableReportAdditionalData = _testConfig.adjust_for_wmf_50;
-                TugAssert.ThrowsExceptionWhen<AggregateException>(
-                        condition: (ex) =>
-                            ex.InnerException is HttpRequestException
-                            && ex.InnerException.Message.Contains(
-                                    "Response status code does not indicate success: 400 (Bad Request)"),
-                        action: () =>
-                            client.SendReport(null).Wait(),
-                        message:
-                            "Throws HTTP exception for bad request (400)");
-            }
-        }
-
-        [TestMethod]
-        public void TestSendReport_BadMissingJobId()
-        {
-            var config = BuildConfig();
-            using (var client = new DscPullClient(config))
-            {
-                client.DisableReportAdditionalData = _testConfig.adjust_for_wmf_50;
-                TugAssert.ThrowsExceptionWhen<AggregateException>(
-                        condition: (ex) =>
-                            ex.InnerException is HttpRequestException
-                            && ex.InnerException.Message.Contains(
-                                    "Response status code does not indicate success: 400 (Bad Request)"),
-                        action: () =>
-                            client.SendReport(new BadSendReportBody()).Wait(),
-                        message:
-                            "Throws HTTP exception for bad request (400)");
-            }
-        }
-
-        [TestMethod]
-        public void TestSendReport_BadDateFormat()
-        {
-            var report = new Model.SendReportBody
-            {
-                JobId = Guid.NewGuid(),
-                StartTime = "NOW",
-                EndTime = "THEN",
-                OperationType = "FOO",
-                ReportFormatVersion = "BAR",
-            };
-            
-            var config = BuildConfig();
-            using (var client = new DscPullClient(config))
-            {
-                client.DisableReportAdditionalData = _testConfig.adjust_for_wmf_50;
-                TugAssert.ThrowsExceptionWhen<AggregateException>(
-                        condition: (ex) =>
-                            ex.InnerException is HttpRequestException
-                            && ex.InnerException.Message.Contains(
-                                    "Response status code does not indicate success: 500 (Internal Server Error)"),
-                        action: () =>
-                            client.SendReport(report).Wait(),
-                        message:
-                            "Throws HTTP exception for internal server error (500)");
-            }
-        }
-
-        [TestMethod]
-        public void TestGetReports_Single()
-        {
-            var reportDateFormat = CLASSIC_SERVER_REPORT_DATE_FORMAT;
-            if (_testConfig.adjust_for_wmf_50)
-                reportDateFormat = CLASSIC_SERVER_REPORT_DATE_FORMAT_ALT;
-
-            var config = BuildConfig(newAgentId: true);
-            var report = new Model.SendReportBody
-            {
-                JobId = Guid.NewGuid(),
-                OperationType = "FOO",
-                RefreshMode = Model.DscRefreshMode.Pull,
-                Status = "BAR",
-                ReportFormatVersion = "Spooky",
-                ConfigurationVersion = "Scary",
-              //StartTime = DateTime.Now.ToString(Model.SendReportBody.REPORT_DATE_FORMAT),
-                StartTime = DateTime.Now.ToString(reportDateFormat),
-              //EndTime = DateTime.Now.ToString(Model.SendReportBody.REPORT_DATE_FORMAT),
-                EndTime = DateTime.Now.ToString(reportDateFormat),
-                RebootRequested = Model.DscTrueFalse.False,
-                StatusData = new[] { "STATUS-DATA" },
-                Errors = new[] { "ERRORS" },
-                AdditionalData = new[]
-                {
-                    new Model.SendReportBody.AdditionalDataItem { Key = "1", Value = "ONE", },
-                    new Model.SendReportBody.AdditionalDataItem { Key = "2", Value = "TWO", },
-                },
-            };
-
-            using (var client = new DscPullClient(config))
-            {
-                client.DisableReportAdditionalData = _testConfig.adjust_for_wmf_50;
-                client.RegisterDscAgent().Wait();
-                client.SendReport(report).Wait();
-
-                var sr = client.GetReports().Result;
-                Assert.IsNotNull(sr, "Reports not null");
-                var srArr = sr.ToArray();
-                Assert.AreEqual(1, srArr.Length, "Reports length is exactly 1");
-
-                // Unfortunate kludge to deal with broken DscService on WMF 5.1
-                //    See https://github.com/PowerShell/PowerShell/issues/2921
-                if (_testConfig.adjust_for_wmf_50)
-                    report.AdditionalData = Model.SendReportBody.AdditionalDataItem.EMPTY_ITEMS;
-
-                var ser1 = JsonConvert.SerializeObject(report);
-                var ser2 = JsonConvert.SerializeObject(srArr[0]);
-                Assert.AreEqual(ser1, ser2, "Submitted and retrieved reports are the same");
-
-                sr = client.GetReports().Result;
-                Assert.IsNotNull(sr, "All reports not null");
-                srArr = sr.ToArray();
-                Assert.AreEqual(1, srArr.Length, "All reports length is exactly 1");
-            }
-        }
-
-        [TestMethod]
-        public void TestGetReports_Multi()
-        {
-            var strArr1 = new[] { "STATUS-1" };
-            var strArr2 = new[] { "STATUS-2" };
-            var strArr3 = new[] { "ERROR-1" };
-            var strArr4 = new[] { "ERROR-2" };
-
-            var config = BuildConfig(newAgentId: true);
-            using (var client = new DscPullClient(config))
-            {
-                client.DisableReportAdditionalData = _testConfig.adjust_for_wmf_50;
-                client.RegisterDscAgent().Wait();
-                client.SendReport(operationType: "1", statusData: strArr1).Wait();
-                client.SendReport(operationType: "2", statusData: strArr2).Wait();
-                client.SendReport(operationType: "3", errors: strArr3).Wait();
-                client.SendReport(operationType: "4", errors: strArr4).Wait();
-
-                var sr = client.GetReports().Result;
-                Assert.IsNotNull(sr, "All reports not null");
-                var srArr = sr.ToArray();
-                Assert.AreEqual(4, srArr.Length, "All reports length");
-
-                var srArrOrd = srArr.OrderBy(x => x.OperationType).ToArray();
-                CollectionAssert.AreEqual(strArr1, srArrOrd[0].StatusData);
-                CollectionAssert.AreEqual(strArr2, srArrOrd[1].StatusData);
-                CollectionAssert.AreEqual(strArr3, srArrOrd[2].Errors);
-                CollectionAssert.AreEqual(strArr4, srArrOrd[3].Errors);
-            }
-        }
-
-        private static DscPullConfig BuildConfig(bool newAgentId = false)
-        {
-            var config = new DscPullConfig();
-
-            config.AgentId = newAgentId
-                    ? Guid.NewGuid()
-                    : Guid.Parse(_testConfig.agent_id);
-
-            config.AgentInformation = Program.ComputeAgentInformation();
-            config.ConfigurationNames = new[] { "TestConfig1" };
-
-            // This is required for RegKey Authz
-            config.CertificateInformation = new Model.CertificateInformation
-            {
-                FriendlyName = "Tug.Client-Test",
-                Issuer = "Tug.Client.Test",
-                NotAfter = DateTime.Now.AddYears(1).ToString("O"),
-                NotBefore = DateTime.Now.AddMinutes(-10).ToString("O"),
-                Subject = "Tug.Client-TestNode",
-                PublicKey = "U3lzdGVtLlNlY3VyaXR5LkNyeXB0b2dyYXBoeS5YNTA5Q2VydGlmaWNhdGVzLlB1YmxpY0tleQ==",
-                Thumbprint = "8351F16C2B06634279F2C0287B5430452DA1CD94",
-                Version = 3,
-            };
-
-            config.ConfigurationRepositoryServer = new DscPullConfig.ServerConfig
-            {
-                ServerUrl = new Uri(_testConfig.server_url),
-                RegistrationKey = _testConfig.reg_key,
-            };
-
-            // Only for debugging/testing in DEV (i.e. with Fiddler)
-            if (!string.IsNullOrEmpty(_testConfig.proxy_url))
-                config.ConfigurationRepositoryServer.Proxy =
-                        new Util.BasicWebProxy(_testConfig.proxy_url);
-
-            // Resource & Reporting Server endpoint URLs same as Config Server endpoint URL
-            config.ResourceRepositoryServer = config.ConfigurationRepositoryServer;
-            config.ReportServer = config.ConfigurationRepositoryServer;
-
-            config.SendReport = new DscPullConfig.SendReportConfig
-            {
-                CommonDefaults = new Model.SendReportBody
-                {
-                    OperationType = "Consistency",
-                    ReportFormatVersion = "2.0",
-                    StartTime = "%NOW%",
-                    AdditionalData = new[]
-                    {
-                        new Model.SendReportBody.AdditionalDataItem
-                        {
-                            Key = "OSVersion",
-                            Value = "{\"VersionString\":\"Microsoft Windows NT 10.0.14393.0\",\"ServicePack\":\"\",\"Platform\":\"Win32NT\"}"
-                        },
-                        new Model.SendReportBody.AdditionalDataItem
-                        {
-                            Key = "PSVersion",
-                            Value = "{\"CLRVersion\":\"4.0.30319.42000\",\"PSVersion\":\"5.1.14393.576\",\"BuildVersion\":\"10.0.14393.576\"}"
-                        },
-                    }
-                },
-                Profiles = new Dictionary<string, Model.SendReportBody>
-                {
-                    ["SimpleInventoryDefaults"] = new Model.SendReportBody
-                    {
-                        NodeName = "HOST_NAME",
-                        IpAddress = "127.0.0.1;::1",
-                        LCMVersion = "2.0",
-                    },
-                    ["DetailedStatusDefaults"] = new Model.SendReportBody
-                    {
-                        RefreshMode = Model.DscRefreshMode.Pull,
-                        Status = "Success",
-                        EndTime = "%NOW%",
-                        ConfigurationVersion = "2.0",
-                        RebootRequested = Model.DscTrueFalse.False,
-                    },
-                    ["ErrorDefaults"] = new Model.SendReportBody
-                    { },
-                },
-            };
-
-            return config;
-        }
-
 
         public class BadVersionTypeCertInfo : Model.CertificateInformation
         {
@@ -751,7 +438,6 @@ namespace Tug.Client
             { get; set; } = "BAR";
         }
 
-
         public class BadFieldOrderCertInfo : Model.CertificateInformation
         {
             public BadFieldOrderCertInfo(Model.CertificateInformation copyFrom) : base(copyFrom)
@@ -760,36 +446,6 @@ namespace Tug.Client
             // Redefining this field forces it to
             // go to the top of serialization order
             public new int Version { get; set; }
-        }
-
-        public class BadSendReportBody : Model.SendReportBody
-        {
-            public new string JobId
-            { get; set; }
-
-            [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-            public new string OperationType
-            { get; set; }
-
-            [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-            public new string ReportFormatVersion
-            { get; set; }
-
-            [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-            public new string StartTime
-            { get; set; }
-
-            [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-            public new string Errors
-            { get; set; }
-
-            [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-            public new string StatusData
-            { get; set; }
-
-            [JsonProperty(NullValueHandling = NullValueHandling.Ignore)]
-            public new string AdditionalData
-            { get; set; }
         }
     }
 }
